requirements.txt
absl-py==2.1.0
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
binwalk==2.1.0
blinker==1.9.0
catboost==1.2.7
certifi==2025.1.31
charset-normalizer==3.4.1
clarabel==0.10.0
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cvxpy==1.6.2
cycler==0.12.1
datasets==3.6.0
debugpy==1.8.13
decorator==5.2.1
dill==0.3.8
executing==2.2.0
filelock==3.13.1
Flask==3.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.6.0
fsspec==2024.6.1
gast==0.6.0
google-pasta==0.2.0
graphviz==0.20.3
greenlet==3.2.0
grpcio==1.71.0
h5py==3.13.0
hf-xet==1.1.0
huggingface-hub==0.31.1
idna==3.10
imbalanced-learn==0.13.0
immutabledict==4.2.1
ipykernel==6.29.5
ipython==9.0.1
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
Keras-Preprocessing==1.1.2
kiwisolver==1.4.8
libclang==18.1.1
lightgbm==4.6.0
llvmlite==0.44.0
Mako==1.3.10
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
namex==0.0.8
narwhals==1.29.1
nest-asyncio==1.6.0
networkx==3.3
numba==0.61.0
numpy==2.2.5
opencv-python==4.11.0.86
opt_einsum==3.4.0
optbinning==0.20.1
optree==0.14.1
ortools==9.11.4210
osqp==0.6.7.post3
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==5.26.1
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyarrow==20.0.0
pydicom==3.0.1
Pygments==2.19.1
PyJWT==2.10.1
pymystem3==0.2.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.1
qdldl==0.1.7.post5
regex==2024.11.6
requests==2.32.3
rich==13.9.4
ropwr==1.1.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
scs==3.2.7.post2
seaborn==0.13.2
sentencepiece==0.2.0
shap==0.47.0
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
SQLAlchemy==2.0.40
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
